initial implementation
The current implementation is very naïve in that for each
subvolume to send, it sends the entire subvolume with no
consideration on incremental sending.  A deeper look at the btrfs
internals might be needed to implement incremental sending.
Consequently, the current implementation requires sending *much*
more data than the size of FS in order to send all the subvolumes
-- essentially the raw sum of sizes for all subvolumes are sent,
regardless of whether data are shared among themselves.

This script requires
[btrfs-list](https://github.com/speed47/btrfs-list) -- but
essentially the idea is to list all subvolumes in a FS relative to
the FS root.

To avoid constantly asking for elevation, one can run the script
with sudo.